% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_eval_functions.R
\name{dm.test.cnr.w}
\alias{dm.test.cnr.w}
\title{CNR (2017), Wald-statistic}
\usage{
dm.test.cnr.w(d, q = 2, cl = 0.05)
}
\arguments{
\item{d}{loss differential}

\item{q}{number of blocks}

\item{cl}{confidence level (default = 0.05, i.e., 5\%)}

\item{R}{number of randomized test stats}
}
\value{
This function returns a class with the following elements
\itemize{
\item \code{rej} is a T/F value. TRUE (reject), FALSE (accept)
\item \code{stat} is a test statistic
\item \code{pval} is an associated p-value
}
}
\description{
Randomization test based on asymptotic symmetry, Wald-statistic
}
\author{
Minchul Shin
}
