% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Tradeoff.R
\name{Plot_Tradeoff}
\alias{Plot_Tradeoff}
\title{Visualizes the size distortion maximum power loss tradeoff}
\usage{
Plot_Tradeoff(
  data,
  f1 = NULL,
  f2 = NULL,
  y = NULL,
  loss_function = NULL,
  n_sim = 1000,
  m_set = NULL,
  no_m_label = FALSE
)
}
\arguments{
\item{data}{Data frame}

\item{f1}{Column containing forecaster 1's predictions. Should be a string.}

\item{f2}{Column containing forecaster 2's predictions. Should be a string.}

\item{y}{Column containing the realized value for the outcome variable. Should be a string.}

\item{loss_function}{The transformation applied to the forecast error. Defaults to squared error loss. The user supplied function should take two inputs and a scalar output, loss = loss_function(f, y). For example, quadratic loss function would be defined as \code{loss_function=function(f,y){(f-y)^2}}.}

\item{n_sim}{The number of simulations used to generate the ARIMA model. Defaults to 1,000.}

\item{m_set}{The truncation parameter. Defaults to c(1:10, seq( 11, floor(nrow(data)/2), 10)). For a standard long-run variance calculation (for example, using Bartlett kernel), it controls the number of terms used in estimating the autocovariance matrix. It should be a vector of integers with the values of M you would like to plot.}

\item{no_m_label}{TRUE if plot without m labels. Defaults to FALSE.}
}
\value{
A list of length 2. The first element is a ggplot2 object of the size-power tradeoff. The second element is the underlying data used to construct the plot in element 1.
}
\description{
`Plot_Tradeoff` creates a plot to show sensitivity of statistical significance to the choice of bandwidth and how size distortion and maximum power loss vary.
}
\examples{

# A typical example
set.seed(1234)
output = Plot_Tradeoff(
  data = TBILL,
  f1   = "SPFfor_Step1",
  f2   = "NCfor_Step1",
  y    = "Realiz1"
)
output[[1]] # The first element is a ggplot2 object of the size-power tradeoff.
output[[2]] # The second element is the underlying data used to construct the plot in element 1.


# An example with a user supplied M values (with a larger set of M values)
set.seed(1234)
Plot_Tradeoff(
  data = TBILL,
  f1 = "SPFfor_Step1",
  f2 = "NCfor_Step1",
  y  = "Realiz1",
  m_set = c(1:10, seq(from = 11, to = nrow(TBILL) - 20, by = 10))
)

# An example with a user spplied loss function
# Suppose we want to use the mean absolute error as a loss function rather than a quadratic loss function
set.seed(1234)
output = Plot_Tradeoff(
  data = TBILL,
  f1   = "SPFfor_Step1",
  f2   = "NCfor_Step1",
  y    = "Realiz1",
  loss_function = function(f,y){ abs(f-y) }
)

# An example without (f1, f2, y). The function will take the first three columns and use them
set.seed(1234)
tmpdata = TBILL[, c("SPFfor_Step1", "NCfor_Step1", "Realiz1")] # data with [f1, f2, y]
Plot_Tradeoff(
  data = tmpdata
)

}
