% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TBILL}
\alias{TBILL}
\title{3-Month Treasury Bill Rate (TBILL)}
\format{
`TBILL`
A data frame with 219 rows and 16 columns
}
\source{
<https://www.philadelphiafed.org/-/media/frbp/assets/surveys-and-data/survey-of-professional-forecasters/data-files/tbill/data_spf_error_statistics_tbill_1_aic.xls?la=en&hash=F432350F84B0E4CCE9A1E9D399447CA9>
}
\usage{
TBILL
}
\description{
Error Statistics for the Survey of Professional Forecasters for Treasury Bill Rate (Three Month). All NAs are replaced with 0 as in Coroneo and Iacone (2020).
}
\references{
Coroneo, L., Iacone, F. (2020), Comparing Predictive Accuracy in Small Samples Using Fixed-Smoothing Asymptotics, \emph{Journal of Applied Econometrics}, \bold{35}(4), 391-409.
}
\keyword{datasets}
