% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UNEMP}
\alias{UNEMP}
\title{Civilian Unemployment Rate (UNEMP)}
\format{
`UNEMP`
A data frame with 219 rows and 16 columns
}
\source{
<https://www.philadelphiafed.org/-/media/frbp/assets/surveys-and-data/survey-of-professional-forecasters/data-files/unemp/data_spf_error_statistics_unemp_1_aic.xls?la=en&hash=4CAD0B11FEAB6C4D0F30C38965FE3354>
}
\usage{
UNEMP
}
\description{
Error Statistics for the Survey of Professional Forecasters for Unemployment Rate. All NAs are replaced with 0 as in Coroneo and Iacone (2020).
}
\references{
Coroneo, L., Iacone, F. (2020), Comparing Predictive Accuracy in Small Samples Using Fixed-Smoothing Asymptotics, \emph{Journal of Applied Econometrics}, \bold{35}(4), 391-409.
}
\keyword{datasets}
