% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Tradeoff.R
\name{Plot_Tradeoff}
\alias{Plot_Tradeoff}
\title{Visualizes the size distortion maximum power loss tradeoff}
\usage{
Plot_Tradeoff(
  data,
  f1 = NULL,
  f2 = NULL,
  y = NULL,
  loss_function = NULL,
  n_sim = 1000,
  m_set = c(1:10, seq(11, nrow(data) - 1, 10))
)
}
\arguments{
\item{f1}{Column containing forecaster 1's predictions. Should be a string.}

\item{f2}{Column containing forecaster 2's predictions. Should be a string.}

\item{y}{Column containing the realized value for the outcome variable. Should be a string.}

\item{loss_function}{The transformation applied to the forecast error. Defaults to squared error loss. Should be a string.}

\item{n_sim}{The number of simulations used to generate the ARIMA model. Defaults to 1,000. Should be a string.}

\item{m_set}{The truncation parameter that controls the number of terms used in estimating the autocovariance matrix. Defaults to M = c(1:10, seq(11, nrow(data) - 1, 10)). Should be a vector of integers with the values of M you would like to plot.}
}
\value{
A list of length 2. The first element is a ggplot2 object of the size-power tradeoff. The second element is the underlying data used to construct the plot in element 1.
}
\description{
`Plot_Tradeoff` creates a plot to show sensitivity of statistical significance to the choice of bandwidth and how size distortion and maximum power loss vary.
}
\examples{

Plot_Tradeoff(
  data = UNEMP,
  f1 = "SPFfor_Step1",
  f2 = "NCfor_Step1",
  y = "Realiz1",
  loss_function = NULL,
  n_sim = 1000,
  m_set = c(1:10, seq(from = 11, to = nrow(data) - 1, by = 10))
)



}
